ML/requirements.txt
Flask==1.0.2
Flask-Cors==3.0.6
newspaper3k==0.2.8
nltk==3.2.5
scikit-learn==0.19.1
scipy==1.1.0
textblob==0.15.3
gunicorn==19.9.0